fix: refactor to avoid circular imports
