+++ src/finchlite/adv_autoschedulers/tensor_stats/__init__.py
@@ -1,27 +1,27 @@
 from ...finch_logic import TensorStats
 from .blocked_stats import BlockedStats, BlockedStatsFactory
-from .database_stats import DatabaseStats, DatabaseStatsFactory
-from .dc_stats import DC, DCStats, DCStatsFactory
 from .dense_stat import DenseStats, DenseStatsFactory
 from .dummy_stats import DummyStats, DummyStatsFactory
+from .numeric_stats import NumericStats
 from .stats_interpreter import StatsInterpreter
 from .tensor_def import TensorDef
 from .tensor_stats import BaseTensorStats
 from .uniform_stats import UniformStats, UniformStatsFactory
 
+# dc_stats and database_stats are NOT imported here because they
+# statically depend on compile/, which must be fully loaded first.
+# Import DC, DCStats, DCStatsFactory, DatabaseStats, DatabaseStatsFactory
+# directly from their submodules when needed.
+
 __all__ = [
-    "DC",
     "BaseTensorStats",
     "BlockedStats",
     "BlockedStatsFactory",
-    "DCStats",
-    "DCStatsFactory",
-    "DatabaseStats",
-    "DatabaseStatsFactory",
     "DenseStats",
     "DenseStatsFactory",
     "DummyStats",
     "DummyStatsFactory",
+    "NumericStats",
     "StatsInterpreter",
     "TensorDef",
     "TensorStats",
