+++ tests/test_cache_loader.py
@@ -3,14 +3,17 @@ import logging
 import numpy as np
 
 import finchlite as fl
+from finchlite.adv_autoschedulers.tensor_stats import (
+    DenseStatsFactory,
+    UniformStatsFactory,
+)
+from finchlite.adv_autoschedulers.tensor_stats.blocked_stats import BlockedStatsFactory
+from finchlite.adv_autoschedulers.tensor_stats.dc_stats import DCStatsFactory
 from finchlite.algebra import ffuncs
 from finchlite.autoschedule.cache import (
     LogicCacheLRU_Embeddings_Norms,
 )
 from finchlite.autoschedule.executor import LogicExecutor
-from finchlite.autoschedule.tensor_stats import DenseStatsFactory, UniformStatsFactory
-from finchlite.autoschedule.tensor_stats.blocked_stats import BlockedStatsFactory
-from finchlite.autoschedule.tensor_stats.dc_stats import DCStatsFactory
 from finchlite.finch_logic import (
     Alias,
     Field,
